# docs/swagger.yaml
# =====================================
# FILE: docs/swagger.yaml
# =====================================
openapi: 3.0.0
info:
  title: Security Demo API
  version: 1.0.0
  description: API documentation for the Web Security Demonstration project
  contact:
    name: API Support
    email: support@securitydemo.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://your-production-url.com/api
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Admin, Moderator, User, Guest]
        created_at:
          type: string
          format: date-time
    
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - captchaAnswer
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        captchaAnswer:
          type: string
    
    Error:
      type: object
      properties:
        error:
          type: string
    
    SecurityTestRequest:
      type: object
      required:
        - input
        - protection
      properties:
        input:
          type: string
        protection:
          type: boolean

paths:
  /health:
    get:
      summary: Health check endpoint
      tags: [System]
      responses:
        200:
          description: System is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string

  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: User registered successfully
        400:
          description: Invalid input
        409:
          description: User already exists
  
  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  csrfToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        401:
          description: Invalid credentials
        429:
          description: Too many requests
  
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Token refreshed
        403:
          description: Invalid refresh token
  
  /auth/logout:
    post:
      summary: Logout user
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Logout successful
        401:
          description: Unauthorized
  
  /users/profile:
    get:
      summary: Get user profile
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
  
  /users:
    get:
      summary: Get all users (Admin/Moderator only)
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
        403:
          description: Forbidden
  
  /activity:
    get:
      summary: Get activity logs
      tags: [Monitoring]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Activity logs
        401:
          description: Unauthorized
  
  /stats:
    get:
      summary: Get security statistics
      tags: [Monitoring]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Security statistics
        401:
          description: Unauthorized
  
  /security-events:
    get:
      summary: Get security events (Admin only)
      tags: [Monitoring]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Security events
        401:
          description: Unauthorized
        403:
          description: Forbidden
  
  /demo/xss:
    post:
      summary: Test XSS vulnerability
      tags: [Security Demo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityTestRequest'
      responses:
        200:
          description: XSS test result
  
  /demo/sql:
    post:
      summary: Test SQL injection vulnerability
      tags: [Security Demo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityTestRequest'
      responses:
        200:
          description: SQL injection test result
  
  /demo/csrf:
    post:
      summary: Test CSRF vulnerability
      tags: [Security Demo]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                csrfToken:
                  type: string
                protection:
                  type: boolean
      responses:
        200:
          description: CSRF test result
        403:
          description: CSRF token invalid
  
  /demo/path-traversal:
    post:
      summary: Test path traversal vulnerability
      tags: [Security Demo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                protection:
                  type: boolean
      responses:
        200:
          description: Path traversal test result
        403:
          description: Path traversal blocked

tags:
  - name: System
    description: System health and status
  - name: Authentication
    description: User authentication endpoints
  - name: Users
    description: User management endpoints
  - name: Monitoring
    description: Activity and security monitoring
  - name: Security Demo
    description: Security vulnerability demonstration endpoints